Added in the high level navmesh scripts
Why is this not just in unity bu default, everything else seems to be working with the plugin now, mainly just need to do documentation. The basic, chase, advanced AI are all working, there is buttons to generate navmesh and rebake it along with create a player. There is another window to make waypoints, that will branch off as well